Move MenuLinkTreeHandler into service folder, cleanup user fields on disabling of functionality
