+++ menu_item_extras.services.yml
@@ -1,6 +1,6 @@
 services:
   menu_item_extras.menu_link_tree_handler:
-    class: Drupal\menu_item_extras\MenuLinkTreeHandler
+    class: Drupal\menu_item_extras\Service\MenuLinkTreeHandler
     arguments: ['@entity_type.manager']
   menu_item_extras.menu_link_content_helper:
     class: Drupal\menu_item_extras\Service\MenuLinkContentService
